fix(backend): fix trying to remove an undefined archive

+++ backend/src/models/userModel.js
@@ -259,9 +259,11 @@ userSchema.methods.sendLoginCode = async function () {
 };
 
 userSchema.methods.deleteArchive = async function () {
-  await fs.remove(this.dicomDataPath);
-  this.dicomDataPath = undefined;
-  await this.save();
+  if (this.dicomDataPath !== undefined) {
+    await fs.remove(this.dicomDataPath);
+    this.dicomDataPath = undefined;
+    await this.save();
+  }
 };
 
 export default mongoose.model("User", userSchema);
